requirements.txt
bleak
Pillow
ttkbootstrap
httpx
